requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.6
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==22.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.7.0
XlsxWriter==3.1.9